Refactor functions into separate packages
